requirements.txt
pytest==8.3.5
requests>=2.31.0
mysql-connector-python>=8.4.0
pymongo>=4.6.0
